refactor: extract WeightedMeshControl and SkeletonWeightProcessor from GlrSkeletonVisual
Extract WeightedMeshControl inner class to top-level public class and
skeleton traversal/rendering logic to package-private SkeletonWeightProcessor.
Deduplicate GL state management with setMeshGlState/resetMeshGlState helpers.
Update UtilityWeightedMeshControl and UtilitySkeletonVisualAdapter imports.

GlrSkeletonVisual reduced from 712 to 448 lines. Public API unchanged.

Closes #616
